feat: add authentication routes and implement sign-in and registration forms with validation
