+++ src/app/auth/login/page.tsx
@@ -1,0 +1,94 @@
+'use client';
+
+import ICON_SET from '@/constants/icons';
+import { APP_ROUTES } from '@/constants/routes.constants';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { Icon } from '@iconify/react';
+import Link from 'next/link';
+import { useState } from 'react';
+import { useForm } from 'react-hook-form';
+import { z } from 'zod';
+
+const signInSchema = z.object({
+    email: z.string().email({ message: 'Invalid email address' }),
+    password: z.string().min(6, { message: 'Password must be at least 6 characters' }),
+});
+
+type SignInFormData = z.infer<typeof signInSchema>;
+
+export default function SignInForm() {
+    const [showPassword, setShowPassword] = useState(false);
+    const {
+        register,
+        handleSubmit,
+        formState: { errors },
+    } = useForm<SignInFormData>({
+        resolver: zodResolver(signInSchema),
+    });
+
+    const onSubmit = (data: SignInFormData) => {
+        console.log('Sign In Data:', data);
+    };
+
+    return (
+        <>
+            <h2 className="text-text-primary mt-2 text-2xl font-semibold">
+                <Icon icon={ICON_SET.EYE} className="mr-2 inline size-7" />
+                Welcome back
+            </h2>
+            <p className="text-text-secondary text-sm">Please enter your details to sign in</p>
+
+            <hr className="my-4" />
+
+            <form onSubmit={handleSubmit(onSubmit)}>
+                <div className="form-group mb-3">
+                    <label className="form-text">Email address</label>
+                    <input type="email" {...register('email')} className="form-field" placeholder="Enter your email" />
+                    {errors.email && <p className="text-xs text-red-500">{errors.email.message}</p>}
+                </div>
+
+                <div className="form-group mb-3">
+                    <label className="form-text">Password</label>
+                    <div className="relative">
+                        <input
+                            type={showPassword ? 'text' : 'password'}
+                            {...register('password')}
+                            className="form-field"
+                            placeholder="Enter your password"
+                        />
+                        <button
+                            type="button"
+                            className="text-text-secondary hover:text-text-primary absolute inset-y-0 right-3 flex cursor-pointer items-center"
+                            onClick={() => setShowPassword((prev) => !prev)}>
+                            <Icon icon={showPassword ? ICON_SET.EYE : ICON_SET.EYE_CLOSE} className="size-6" />
+                        </button>
+                    </div>
+                    {errors.password && <p className="text-xs text-red-500">{errors.password.message}</p>}
+                </div>
+
+                <div className="flex items-center justify-between">
+                    <label className="form-checkbox">
+                        <input type="checkbox" className="checkbox-field" />
+                        <p className="form-text"> Remember for 30 days</p>
+                    </label>
+                    <Link href={APP_ROUTES.AUTH.FORGOT_PASSWORD} className="text-highlight text-xs hover:underline">
+                        Forgot password?
+                    </Link>
+                </div>
+
+                <hr className="my-6" />
+
+                <button type="submit" className="button w-full">
+                    Sign in
+                </button>
+            </form>
+
+            <p className="text-text-secondary mt-2 text-center text-sm">
+                Don’t have an account?{' '}
+                <Link href={APP_ROUTES.AUTH.REGISTER} className="text-highlight hover:underline">
+                    Create account
+                </Link>
+            </p>
+        </>
+    );
+}
