+++ src/constants/icons.ts
@@ -15,8 +15,8 @@ enum ICON_SET {
     EDIT_PROFILE = 'fluent:person-edit-24-filled',
     EMAIL = 'line-md:email-alt-filled',
     ERROR = 'meteocons:code-red-fill',
-    EYE = 'solar:eye-outline',
-    EYE_CLOSE = 'solar:eye-closed-linear',
+    EYE = 'line-md:watch-loop',
+    EYE_CLOSE = 'line-md:watch-off-loop',
     FILE_PLUS = 'line-md:file-plus-filled',
     FILTER = 'line-md:filter-filled',
     GAMEPAD = 'emojione-monotone:video-game',
@@ -36,7 +36,6 @@ enum ICON_SET {
     MOON = 'meteocons:clear-night-fill',
     MORE_DOTS = 'pepicons-pop:dots-y',
     MUSIC = 'solar:music-note-bold',
-    PASSWORD = 'line-md:watch-loop',
     PERSON = 'line-md:person-filled',
     PLAYER = 'game-icons:centurion-helmet',
     PLUS = 'line-md:plus',
@@ -47,7 +46,6 @@ enum ICON_SET {
     SETTINGS = 'solar:settings-bold',
     SETTINGS_OUTLINED = 'solar:settings-outline',
     SHARE = 'solar:share-bold-duotone',
-    SHOW_PASSWORD = 'line-md:watch-off-loop',
     SPOTIFY = 'line-md:spotify-filled',
     SUCCESS = 'line-md:confirm-circle-filled',
     SUN = 'meteocons:clear-day-fill',
